pyproject.toml
[project]
name = "fake-news-detection-main"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.8",
    "joblib>=1.5.3",
    "pillow>=12.0.0",
    "python-multipart>=0.0.22",
    "scikit-learn>=1.8.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.1.0",
    "uvicorn>=0.40.0",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cpu"
